Path to CmdStan for instantiate

Return the directory path to CmdStan that the instantiate package uses.

stan_cmdstan_path(cmdstan_install = Sys.getenv("CMDSTAN_INSTALL"))

Arguments

cmdstan_install

Character of length 1, how CmdStan was installed. The installation method determines how instantiate looks up CmdStan. Set the default value with the CMDSTAN_INSTALL environment variable. Choices:

"internal": Use the copy of CmdStan installed internally inside the instantiate package file system. Before using this option, CmdStan must be installed in the package file system. To do this, set the CMDSTAN_INSTALL environment variable to "internal" before installing instantiate.

"fixed": Use the copy of CmdStan that was located at the value contained in Sys.getenv("CMDSTAN") at the time instantiate was installed. In this case, the path to CmdStan is not inside instantiate itself, but it is fixed at installation time and does not depend on the current value of Sys.getenv("CMDSTAN") at runtime.

"cmdstanr": Let the cmdstanr::cmdstan_path() decide where to look for CmdStan. The cmdstanr package must be installed. If it is not installed, the function returns the empty string "".

"" (default): Try all 3 options in the order above to find a valid installed copy of CmdStan.

Value

Character of length 1, directory path to CmdStan. The path may or may not exist. Returns the empty string "" if the path cannot be determined.

See also

Examples

stan_cmdstan_path()
#> [1] "/home/runner/.cmdstan/cmdstan-2.32.2"