Get a pre-compiled Stan model built into an R package.

Get the cmdstanr model object for a pre-compiled Stan model inside a package.

stan_package_model(
  name,
  package,
  library = NULL,
  cmdstan_install = Sys.getenv("CMDSTAN_INSTALL")
)

Arguments

name

Character of length 1, base name of the Stan model file (without the containing directory or .stan file extension).

package

Character of length 1, name of the R package to look for the built-in Stan model.

library

Character of length 1 or NULL, library path to look for the package with the built-in Stan model. Passed to the lib.loc argument of system.file().

cmdstan_install

Character of length 1, how CmdStan was installed. The installation method determines how instantiate looks up CmdStan. Set the default value with the CMDSTAN_INSTALL environment variable. Choices:

"internal": Use the copy of CmdStan installed internally inside the instantiate package file system. Before using this option, CmdStan must be installed in the package file system. To do this, set the CMDSTAN_INSTALL environment variable to "internal" before installing instantiate.

"fixed": Use the copy of CmdStan that was located at the value contained in Sys.getenv("CMDSTAN") at the time instantiate was installed. In this case, the path to CmdStan is not inside instantiate itself, but it is fixed at installation time and does not depend on the current value of Sys.getenv("CMDSTAN") at runtime.

"cmdstanr": Let the cmdstanr::cmdstan_path() decide where to look for CmdStan. The cmdstanr package must be installed. If it is not installed, the function returns the empty string "".

"" (default): Try all 3 options in the order above to find a valid installed copy of CmdStan.

Value

An R6 Stan model object from the cmdstanr package. Please visit the documentation website at https://mc-stan.org/cmdstanr/

for detailed information on the composition of this model object and how to use it to conduct Bayesian inference.

Details

Packages configured with instantiate compile their Stan models on installation. Then the stan_package_model() function retrieves the cmdstanr::cmdstan_model() object without needing to re-compile the model. Please see the documentation website of the instantiate package for examples.

Examples

# Please see the documentation website of the {instantiate} package
#   for examples.