List Stan model file paths.

List all the paths to the Stan model files in a package.

stan_package_model_files(path = getwd())

Arguments

path

Character of length 1, path to the root directory of the R package with the Stan models.

Value

Character vector of paths to Stan model files in the package.

Details

All Stan models must live in the inst/stan/ directory in the package file system.

See also

Examples

path <- tempfile()
stan_package_create(path = path)
#> Package with an internal Stan model created at /tmp/RtmpWoBUcl/file2b9b32caea2d. Configure with stan_package_configure() before installing.
stan_package_model_files(path)
#> [1] "/tmp/RtmpWoBUcl/file2b9b32caea2d/inst/stan/bernoulli.stan"