Compile the Stan models in an R package.

Compile all Stan models in a directory, usually in a package.

stan_package_compile(
  models = instantiate::stan_package_model_files(),
  cmdstan_install = Sys.getenv("CMDSTAN_INSTALL"),
  verbose = TRUE,
  quiet = FALSE,
  pedantic = FALSE,
  include_paths = NULL,
  user_header = NULL,
  cpp_options = list(),
  stanc_options = list(),
  force_recompile = getOption("cmdstanr_force_recompile", default = FALSE),
  compile_model_methods = FALSE,
  compile_hessian_method = FALSE,
  compile_standalone = FALSE,
  threads = FALSE
)

Arguments

models

Character vector of file paths to Stan model source code files. Defaults to the Stan files in ./inst/stan/ because all the Stan model files must live in the inst/stan/ folder for an R package built with instantiate.

cmdstan_install

Character of length 1, how CmdStan was installed. The installation method determines how instantiate looks up CmdStan. Set the default value with the CMDSTAN_INSTALL environment variable. Choices:

"internal": Use the copy of CmdStan installed internally inside the instantiate package file system. Before using this option, CmdStan must be installed in the package file system. To do this, set the CMDSTAN_INSTALL environment variable to "internal" before installing instantiate.

"fixed": Use the copy of CmdStan that was located at the value contained in Sys.getenv("CMDSTAN") at the time instantiate was installed. In this case, the path to CmdStan is not inside instantiate itself, but it is fixed at installation time and does not depend on the current value of Sys.getenv("CMDSTAN") at runtime.

"cmdstanr": Let the cmdstanr::cmdstan_path() decide where to look for CmdStan. The cmdstanr package must be installed. If it is not installed, the function returns the empty string "".

"" (default): Try all 3 options in the order above to find a valid installed copy of CmdStan.

verbose

Logical of length 1, whether to set the cmdstanr_verbose global option to print more compiler messages than usual.

quiet

Argument to cmdstanr::cmdstan_model() to control compilation.

pedantic

Logical of length 1, whether to activate pedantic mode when compiling the Stan models. See the cmdstanr documentation for details.

include_paths

Argument to cmdstanr::cmdstan_model() to control model compilation.

user_header

Argument to cmdstanr::cmdstan_model() to control model compilation.

cpp_options

Argument to cmdstanr::cmdstan_model() to control model compilation.

stanc_options

Argument to cmdstanr::cmdstan_model() to model control compilation.

force_recompile

Argument to cmdstanr::cmdstan_model() to model control compilation.

compile_model_methods

Argument to cmdstanr::cmdstan_model() to model control compilation.

compile_hessian_method

Argument to cmdstanr::cmdstan_model() to model control compilation.

compile_standalone

Argument to cmdstanr::cmdstan_model() to model control compilation.

threads

Argument to cmdstanr::cmdstan_model() to model control compilation.

Value

NULL (invisibly). Called for its side effects.

Details

If building a package using instantiate, all Stan model files must live in a folder called inst/stan/ in the package source directory.

See also

Examples

if (identical(Sys.getenv("INSTANTIATE_EXAMPLES"), "true")) {
path <- tempfile()
stan_package_create(path = path)
stan_package_configure(path = path)
models <- stan_package_model_files(path)
list.files(file.path(path, "inst", "stan"))
stan_package_compile(models)
list.files(file.path(path, "inst", "stan"))
}